col_pragma_logro_pgm_extraer_tabla_dynamodb/requirements.txt
awsglue-dev==2021.12.30
boto3==1.40.67
botocore==1.40.67
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
exceptiongroup==1.3.0
idna==3.11
iniconfig==2.3.0
Jinja2==3.1.6
jmespath==1.0.1
MarkupSafe==3.0.3
moto==5.1.16
packaging==25.0
pluggy==1.6.0
py4j==0.10.9.7
pycparser==2.23
pyspark==3.5.1
pytest==8.4.2
pytest-mock==3.15.1
python-dateutil==2.9.0.post0
PyYAML==6.0.3
requests==2.32.5
responses==0.25.8
s3transfer==0.14.0
six==1.17.0
tomli==2.3.0
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
xmltodict==1.0.2
